requirements.txt
# Requirements
# https://pip.pypa.io/en/stable/reference/pip_install/#requirements-file-format

PySide2
PyPDF2
python-docx
gdown
diff_match_patch
googletrans
pypinyin
langdetect
PyGithub